Added logic for CompositeWidgetLocator
