+++ org.jboss.reddeer.swt/META-INF/MANIFEST.MF
@@ -35,6 +35,7 @@ Export-Package: org.jboss.reddeer.swt,
  org.jboss.reddeer.swt.impl.text,
  org.jboss.reddeer.swt.impl.toolbar,
  org.jboss.reddeer.swt.impl.tree,
+ org.jboss.reddeer.swt.locate,
  org.jboss.reddeer.swt.lookup.impl,
  org.jboss.reddeer.swt.matcher,
  org.jboss.reddeer.swt.regex,
